Avoid swapping image back on scroll event
Fixed image lazy loading loadImage so it doesn't replace dynamically swapped images on scroll event.  When custom javascript does an intentional swap on the img src attribute of a lazyloaded image, it swaps back to the original when the user scrolls the browser window.  Instead it should be compared to the placeholder.